requirements.txt

aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow==9.5.0
psutil
wget
pytube
py-tgcalls==1.2.9
ntgcalls==1.1.2
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
urllib3
httpx
SafoneAPI
git+https://github.com/yt-dlp/yt-dlp@master
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
git+https://github.com/naya1503/mypyrogram@beta
croniter
tornado